BOAC-210 Split up merged.py
